feat: useCardPreview 생성
